requirements.txt
torch
torchvision
opencv-python
numpy
streamlit
Pillow
diffusers
transformers  # Required for StableDiffusionInpaintPipeline